broke out the rest extensions a bit and beefed up the multipart form functionality
