+++ src/Sharpener/Sharpener.csproj
@@ -4,7 +4,7 @@
     <Nullable>enable</Nullable>
     <LangVersion>11</LangVersion>
     <PackageId>Sharpener</PackageId>
-    <Version>0.1.9</Version>
+    <Version>0.2.0</Version>
     <Authors>Ryan Singleton</Authors>
     <Company>North of Real</Company>
     <PackageProjectUrl>https://github.com/ryan-singleton/Sharpener</PackageProjectUrl>
