+++ test/Sharpener.Rest.Tests/Extensions/AuthenticationExtensionTests.cs
@@ -12,7 +12,7 @@ public class AuthenticationExtensionTests
     public void GetAuthToken_HeaderDictionary_ShouldReturnNull_WhenSchemeDoesNotExist()
     {
         var headers = new DefaultHttpContext().Request.Headers;
-        headers.Add(RestExtensions.AuthHeader, "Bearer testToken");
+        headers.Add(AuthExtensions.AuthHeader, "Bearer testToken");
 
         var token = headers.GetAuthToken("Basic");
 
@@ -23,7 +23,7 @@ public class AuthenticationExtensionTests
     public void GetAuthToken_HeaderDictionary_ShouldReturnToken_WhenSchemeExists()
     {
         var headers = new DefaultHttpContext().Request.Headers;
-        headers.Add(RestExtensions.AuthHeader, "Bearer testToken");
+        headers.Add(AuthExtensions.AuthHeader, "Bearer testToken");
 
         var token = headers.GetAuthToken("Bearer");
 
@@ -34,7 +34,7 @@ public class AuthenticationExtensionTests
     public void GetAuthToken_HttpHeaders_ShouldReturnNull_WhenSchemeDoesNotExist()
     {
         var headers = new HttpRequestMessage().Headers;
-        headers.Add(RestExtensions.AuthHeader, "Bearer testToken");
+        headers.Add(AuthExtensions.AuthHeader, "Bearer testToken");
 
         var token = headers.GetAuthToken("Basic");
 
@@ -45,7 +45,7 @@ public class AuthenticationExtensionTests
     public void GetAuthToken_HttpHeaders_ShouldReturnToken_WhenSchemeExists()
     {
         var headers = new HttpRequestMessage().Headers;
-        headers.Add(RestExtensions.AuthHeader, "Bearer testToken");
+        headers.Add(AuthExtensions.AuthHeader, "Bearer testToken");
 
         var token = headers.GetAuthToken("Bearer");
 
@@ -56,7 +56,7 @@ public class AuthenticationExtensionTests
     public void GetBasicToken_HeaderDictionary_ShouldReturnNull_WhenBasicSchemeDoesNotExist()
     {
         var headers = new DefaultHttpContext().Request.Headers;
-        headers.Add(RestExtensions.AuthHeader, "Bearer testToken");
+        headers.Add(AuthExtensions.AuthHeader, "Bearer testToken");
 
         var token = headers.GetBasicToken();
 
@@ -67,7 +67,7 @@ public class AuthenticationExtensionTests
     public void GetBasicToken_HeaderDictionary_ShouldReturnToken_WhenBasicSchemeExists()
     {
         var headers = new DefaultHttpContext().Request.Headers;
-        headers.Add(RestExtensions.AuthHeader, "Basic dXNlcm5hbWU6cGFzc3dvcmQ=");
+        headers.Add(AuthExtensions.AuthHeader, "Basic dXNlcm5hbWU6cGFzc3dvcmQ=");
 
         var token = headers.GetBasicToken();
 
@@ -78,7 +78,7 @@ public class AuthenticationExtensionTests
     public void GetBasicToken_HttpHeaders_ShouldReturnNull_WhenBasicSchemeDoesNotExist()
     {
         var headers = new HttpRequestMessage().Headers;
-        headers.Add(RestExtensions.AuthHeader, "Bearer testToken");
+        headers.Add(AuthExtensions.AuthHeader, "Bearer testToken");
 
         var token = headers.GetBasicToken();
 
@@ -89,7 +89,7 @@ public class AuthenticationExtensionTests
     public void GetBasicToken_HttpHeaders_ShouldReturnToken_WhenBasicSchemeExists()
     {
         var headers = new HttpRequestMessage().Headers;
-        headers.Add(RestExtensions.AuthHeader, "Basic dXNlcm5hbWU6cGFzc3dvcmQ=");
+        headers.Add(AuthExtensions.AuthHeader, "Basic dXNlcm5hbWU6cGFzc3dvcmQ=");
 
         var token = headers.GetBasicToken();
 
@@ -100,7 +100,7 @@ public class AuthenticationExtensionTests
     public void GetBearerToken_HeaderDictionary_ShouldReturnNull_WhenBearerSchemeDoesNotExist()
     {
         var headers = new DefaultHttpContext().Request.Headers;
-        headers.Add(RestExtensions.AuthHeader, "Basic dXNlcm5hbWU6cGFzc3dvcmQ=");
+        headers.Add(AuthExtensions.AuthHeader, "Basic dXNlcm5hbWU6cGFzc3dvcmQ=");
 
         var token = headers.GetBearerToken();
 
@@ -111,7 +111,7 @@ public class AuthenticationExtensionTests
     public void GetBearerToken_HeaderDictionary_ShouldReturnToken_WhenBearerSchemeExists()
     {
         var headers = new DefaultHttpContext().Request.Headers;
-        headers.Add(RestExtensions.AuthHeader, "Bearer testToken");
+        headers.Add(AuthExtensions.AuthHeader, "Bearer testToken");
 
         var token = headers.GetBearerToken();
 
@@ -122,7 +122,7 @@ public class AuthenticationExtensionTests
     public void GetBearerToken_HttpHeaders_ShouldReturnNull_WhenBearerSchemeDoesNotExist()
     {
         var headers = new HttpRequestMessage().Headers;
-        headers.Add(RestExtensions.AuthHeader, "Basic dXNlcm5hbWU6cGFzc3dvcmQ=");
+        headers.Add(AuthExtensions.AuthHeader, "Basic dXNlcm5hbWU6cGFzc3dvcmQ=");
 
         var token = headers.GetBearerToken();
 
@@ -133,7 +133,7 @@ public class AuthenticationExtensionTests
     public void GetBearerToken_HttpHeaders_ShouldReturnToken_WhenBearerSchemeExists()
     {
         var headers = new HttpRequestMessage().Headers;
-        headers.Add(RestExtensions.AuthHeader, "Bearer testToken");
+        headers.Add(AuthExtensions.AuthHeader, "Bearer testToken");
 
         var token = headers.GetBearerToken();
 
@@ -148,19 +148,19 @@ public class AuthenticationExtensionTests
         headers.SetAuthToken("Bearer", "testToken");
 
         headers.Should().ContainSingle(h =>
-            h.Key == RestExtensions.AuthHeader && h.Value.Contains("Bearer testToken"));
+            h.Key == AuthExtensions.AuthHeader && h.Value.Contains("Bearer testToken"));
     }
 
     [Fact]
     public void SetAuthToken_HeaderDictionary_ShouldUpdateTokenInHeaders()
     {
         var headers = new DefaultHttpContext().Request.Headers;
-        headers.Add(RestExtensions.AuthHeader, "Bearer oldToken");
+        headers.Add(AuthExtensions.AuthHeader, "Bearer oldToken");
 
         headers.SetAuthToken("Bearer", "newToken");
 
         headers.Should().ContainSingle(h =>
-            h.Key == RestExtensions.AuthHeader && h.Value.Contains("Bearer newToken"));
+            h.Key == AuthExtensions.AuthHeader && h.Value.Contains("Bearer newToken"));
     }
 
     [Fact]
@@ -171,19 +171,19 @@ public class AuthenticationExtensionTests
         headers.SetAuthToken("Bearer", "testToken");
 
         headers.Should().ContainSingle(h =>
-            h.Key == RestExtensions.AuthHeader && h.Value.Contains("Bearer testToken"));
+            h.Key == AuthExtensions.AuthHeader && h.Value.Contains("Bearer testToken"));
     }
 
     [Fact]
     public void SetAuthToken_HttpHeaders_ShouldUpdateTokenInHeaders()
     {
         var headers = new HttpRequestMessage().Headers;
-        headers.Add(RestExtensions.AuthHeader, "Bearer oldToken");
+        headers.Add(AuthExtensions.AuthHeader, "Bearer oldToken");
 
         headers.SetAuthToken("Bearer", "newToken");
 
         headers.Should().ContainSingle(h =>
-            h.Key == RestExtensions.AuthHeader && h.Value.Contains("Bearer newToken"));
+            h.Key == AuthExtensions.AuthHeader && h.Value.Contains("Bearer newToken"));
     }
 
     [Fact]
@@ -194,7 +194,7 @@ public class AuthenticationExtensionTests
         headers.SetBasicToken("username", "password");
 
         headers.Should()
-            .ContainSingle(h => h.Key == RestExtensions.AuthHeader && h.Value[0]!.Contains("Basic "));
+            .ContainSingle(h => h.Key == AuthExtensions.AuthHeader && h.Value[0]!.Contains("Basic "));
         headers.GetBasicToken().Should().NotBeNullOrWhiteSpace();
     }
 
@@ -206,7 +206,7 @@ public class AuthenticationExtensionTests
         headers.SetBasicToken("username", "password");
 
         headers.Should().ContainSingle(h =>
-            h.Key == RestExtensions.AuthHeader && h.Value.First().Contains("Basic "));
+            h.Key == AuthExtensions.AuthHeader && h.Value.First().Contains("Basic "));
         headers.GetBasicToken().Should().NotBeNullOrWhiteSpace();
     }
 
@@ -218,7 +218,7 @@ public class AuthenticationExtensionTests
         headers.SetBearerToken("testToken");
 
         headers.Should().ContainSingle(h =>
-            h.Key == RestExtensions.AuthHeader && h.Value.Contains("Bearer testToken"));
+            h.Key == AuthExtensions.AuthHeader && h.Value.Contains("Bearer testToken"));
     }
 
     [Fact]
@@ -229,6 +229,6 @@ public class AuthenticationExtensionTests
         headers.SetBearerToken("testToken");
 
         headers.Should().ContainSingle(h =>
-            h.Key == RestExtensions.AuthHeader && h.Value.Contains("Bearer testToken"));
+            h.Key == AuthExtensions.AuthHeader && h.Value.Contains("Bearer testToken"));
     }
 }
